+++ test/functions_test.py
@@ -228,7 +228,7 @@ True
 >>> _ = parser.add_argument('--version', help='version', nargs="?", const=1)
 >>> args = parser.parse_args({"--version":True})
 >>> run(args)
-1.0
+1.1
 >>>
 >>> args = parser.parse_args(["--method", 'v60'])
 >>> run(args)