+++ CHANGELOG.md
@@ -6,12 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- 2 new water units
+- 3 new water units
 	1. Troy Ounce (`t oz`)
 	2. Grain (`gr`)
-- 2 new coffee units
+	3. Carats (`ct`)
+- 3 new coffee units
 	1. Troy Ounce (`t oz`)
 	2. Grain (`gr`)
+	3. Carats (`ct`)
 ## [1.0] - 2024-12-17
 ### Added
 - 3 new water units