+++ CHANGELOG.md
@@ -6,10 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- 3 new water units
+- 4 new water units
 	1. Troy Ounce (`t oz`)
 	2. Grain (`gr`)
 	3. Carats (`ct`)
+	4. Cubic Centimeters (`cc`)
 - 3 new coffee units
 	1. Troy Ounce (`t oz`)
 	2. Grain (`gr`)