+++ CHANGELOG.md
@@ -6,11 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- 4 new water units
+- 5 new water units
 	1. Troy Ounce (`t oz`)
 	2. Grain (`gr`)
 	3. Carats (`ct`)
 	4. Cubic Centimeters (`cc`)
+	5. Centiliters (`cl`)
 - 3 new coffee units
 	1. Troy Ounce (`t oz`)
 	2. Grain (`gr`)