+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	1. Troy Ounce (`t oz`)
 	2. Grain (`gr`)
 	3. Carats (`ct`)
+### Changed
+- `README.md` updated
 ## [1.0] - 2024-12-17
 ### Added
 - 3 new water units