+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.1] - 2025-01-02
 ### Added
 - 5 new water units
 	1. Troy Ounce (`t oz`)
@@ -125,7 +126,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	5. Siphon
 	6. Custom
 
-[Unreleased]: https://github.com/sepandhaghighi/mycoffee/compare/v1.0...dev
+[Unreleased]: https://github.com/sepandhaghighi/mycoffee/compare/v1.1...dev
+[1.1]: https://github.com/sepandhaghighi/mycoffee/compare/v1.0...v1.1
 [1.0]: https://github.com/sepandhaghighi/mycoffee/compare/v0.9...v1.0
 [0.9]: https://github.com/sepandhaghighi/mycoffee/compare/v0.8...v0.9
 [0.8]: https://github.com/sepandhaghighi/mycoffee/compare/v0.7...v0.8